Fix external build files

+++ externals/coda-oss/build/makewheel.py
@@ -1,4 +1,5 @@
 from waflib.Build import BuildContext
+from waflib import Errors
 
 import glob
 import os
@@ -16,10 +17,18 @@ class makewheel(BuildContext):
         if not self.all_envs:
             self.load_envs()
 
+        if not self.env['SETUP_PY_DIR']:
+            raise Errors.WafError('Could not make wheel. '
+                'Unable to find setup.py')
         shutil.copyfile(os.path.join(self.env['SETUP_PY_DIR'],
             'setup.py'), 'setup.py')
         self.to_log('Creating wheel\n')
         subprocess.call(['pip', 'wheel', '.', '--wheel-dir', '.', '--no-deps'])
         os.remove('setup.py')
+
+        wheels = glob.glob('*.whl')
+        for wheel in wheels:
+            shutil.copyfile(os.path.join(self.env['install_bindir'], wheel),
+                wheel)
